+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Released: [Place date of release here]
 ### New Features
 
 `faspex5`: New commands `admin file_processing` and `packages file_processing`.
+`aoc`: Command `user settings` now support `create`, `delete`, `list`, `show`, `modify`.
 
 ### Issues Fixed