+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Released: [Place date of release here]
 
 `faspex5`: New commands `admin file_processing` and `packages file_processing`.
 `aoc`: Command `user settings` now support `create`, `delete`, `list`, `show`, `modify`.
+`aoc`: New command `network_policy` and `configuration_policy` in `admin`.
 
 ### Issues Fixed